feature: update barchart
